Ajustes em nomes de pastas e criação de classes para análise do número de ciclos num grafo
